BulletinBoard/BulletinBoard/requirements.txt
APScheduler==3.9.1.post1
asgiref==3.5.2
backports.zoneinfo==0.2.1
certifi==2022.9.24
cffi==1.15.1
charset-normalizer==2.1.1
cryptography==38.0.1
defusedxml==0.7.1
Django==4.1.2
django-allauth==0.51.0
django-apscheduler==0.6.2
django-ckeditor==6.5.1
django-ckeditor-youtube-plugin==0.0.1
django-js-asset==2.0.0
idna==3.4
oauthlib==3.2.2
Pillow==9.3.0
pycparser==2.21
PyJWT==2.6.0
python-dotenv==0.21.0
python3-openid==3.2.0
pytz==2022.6
pytz-deprecation-shim==0.1.0.post0
requests==2.28.1
requests-oauthlib==1.3.1
six==1.16.0
sqlparse==0.4.3
tzdata==2022.5
tzlocal==4.2
urllib3==1.26.12
